correção de exceções e inclusão de classe global de tratamento

+++ src/main/java/com/javanautas/fakeapius/business/service/ProdutoService.java
@@ -3,11 +3,15 @@ package com.javanautas.fakeapius.business.service;
 import com.javanautas.fakeapius.apiv1.dto.ProductsDTO;
 import com.javanautas.fakeapius.business.converter.ProdutoConverter;
 import com.javanautas.fakeapius.infrastructure.entities.ProdutoEntity;
+import com.javanautas.fakeapius.infrastructure.exceptions.BusinessException;
+import com.javanautas.fakeapius.infrastructure.exceptions.ConflictException;
+import com.javanautas.fakeapius.infrastructure.exceptions.UnprocessableEntityException;
 import com.javanautas.fakeapius.infrastructure.repositories.ProdutoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 import static java.lang.String.format;
 
@@ -22,25 +26,37 @@ public class ProdutoService {
         try{
             return repository.save(entity);
         } catch (Exception e) {
-            throw new RuntimeException("Erro ao salvar Produtos" + e);
+            throw new BusinessException("Erro ao salvar Produtos" + e);
         }
     }
 
     public ProductsDTO salvaProdutoDTO(ProductsDTO dto) {
         try{
+            Boolean retorno = existsPorNome(dto.getNome());
+            if(retorno.equals(true)) {
+                throw new ConflictException("Produto já existente no banco de dados " + dto.getNome());
+            }
             ProdutoEntity entity = converter.toEntity(dto);
             return converter.toDTO(repository.save(entity));
+        }catch (ConflictException e) {
+            throw new ConflictException(e.getMessage());
         } catch (Exception e) {
-            throw new RuntimeException("Erro ao salvar Produtos" + e);
+            throw new BusinessException("Erro ao salvar Produtos" + e);
         }
     }
 
 
     public ProductsDTO buscaProdutoPorNome(String nome) {
         try {
-            return converter.toDTO(repository.findByNome(nome));
+            ProdutoEntity produto = repository.findByNome(nome);
+            if (Objects.isNull(produto)) {
+                throw new UnprocessableEntityException("Não foram encontrados produtos com o nome " + nome);
+            }
+            return converter.toDTO(produto);
+        }catch (UnprocessableEntityException e) {
+            throw new UnprocessableEntityException(e.getMessage());
         }catch (Exception e) {
-            throw new RuntimeException(format("Erro ao buscar produto por nome", nome), e);
+            throw new BusinessException(format("Erro ao buscar produto por nome = %s", nome), e);
         }
     }
 
@@ -48,15 +64,23 @@ public class ProdutoService {
         try {
             return converter.toListDTO(repository.findAll());
         }catch (Exception e) {
-            throw new RuntimeException(format("Erro ao buscar todos os produtos"), e);
+            throw new BusinessException("Erro ao buscar todos os produtos", e);
         }
     }
 
     public void deletaProduto(String nome){
         try{
-            repository.deleteByNome(nome);
+            Boolean retorno = existsPorNome(nome);
+            if(retorno.equals(false)) {
+                throw new UnprocessableEntityException("Não foi possível deletar o produto, pois não existe produto com o nome " + nome);
+            }
+            else {
+                repository.deleteByNome(nome);
+            }
+        }catch (UnprocessableEntityException e) {
+            throw new UnprocessableEntityException(e.getMessage());
         }catch (Exception e) {
-            throw new RuntimeException(format("Erro ao deletar produto por nome", nome), e);
+            throw new BusinessException(format("Erro ao deletar produto por nome %s", nome), e);
         }
     }
 
@@ -64,17 +88,19 @@ public class ProdutoService {
         try {
             return repository.existsByNome(nome);
         }catch (Exception e) {
-            throw new RuntimeException(format("Erro ao buscar produto por nome", nome), e);
+            throw new BusinessException(format("Erro ao buscar produto por nome %s", nome), e);
         }
     }
 
     public ProductsDTO updateProduto(String id, ProductsDTO dto){
         try{
-            ProdutoEntity entity = repository.findById(id).orElseThrow(() -> new RuntimeException("Id não existe no banco de dados"));
+            ProdutoEntity entity = repository.findById(id).orElseThrow(() -> new UnprocessableEntityException("Produto não encontrado na base de dados"));
             salvaProdutos(converter.toEntityUpdate(entity, dto, id));
             return converter.toDTO(repository.findByNome(entity.getNome()));
+        }catch (UnprocessableEntityException e) {
+            throw new UnprocessableEntityException(e.getMessage());
         }catch (Exception e) {
-            throw new RuntimeException(format("Erro ao atualizar produto"), e);
+            throw new BusinessException("Erro ao atualizar produto", e);
         }
     }
 }
